Make sure to not override existing mountable file

+++ group_composition_plot/create_volume.sh
@@ -1,3 +1,8 @@
-dd if=/dev/zero of=submissions.volume bs=1024 count=10240
-mkfs.ext4 submissions.volume
-mkdir submissions
+if [ -f submissions.volume ]; then
+    echo "submissions.volume already exists"
+    exit 1
+else
+    dd if=/dev/zero of=submissions.volume bs=1024 count=10240
+    mkfs.ext4 submissions.volume
+fi
+mkdir -p submissions
